Move plot_utils to plot_commands
