+++ freqtrade/commands/plot_commands.py
@@ -2,7 +2,7 @@ from typing import Any, Dict
 
 from freqtrade.exceptions import OperationalException
 from freqtrade.state import RunMode
-from freqtrade.commands import setup_utils_configuration
+from freqtrade.commands.utils import setup_utils_configuration
 
 
 def validate_plot_args(args: Dict[str, Any]):
