+++ tests/test_plotting.py
@@ -11,7 +11,7 @@ from freqtrade.configuration import TimeRange
 from freqtrade.data import history
 from freqtrade.data.btanalysis import create_cum_profit, load_backtest_data
 from freqtrade.exceptions import OperationalException
-from freqtrade.plot.plot_utils import start_plot_dataframe, start_plot_profit
+from freqtrade.commands import start_plot_dataframe, start_plot_profit
 from freqtrade.plot.plotting import (add_indicators, add_profit,
                                      create_plotconfig,
                                      generate_candlestick_graph,
